Moved Debounce into a util file
